+++ src/parse/bitvec.h
@@ -7,13 +7,13 @@
 
 typedef struct bv64 {
     uint32_t width;
-    int64_t value;
+    uint64_t value;
 } bv64_t;
 
 // Convert a string of '0's and '1's to a constant
-int bv64_from_bin_str(char *str, uint32_t n, bv64_t *ans);
+int bv64_from_bin_str(char *str, size_t n, bv64_t *ans);
 
 // Convert a string of hex digits to a constant
-int bv64_from_hex_str(char *str, uint32_t n, bv64_t *ans);
+int bv64_from_hex_str(char *str, size_t n, bv64_t *ans);
 
 #endif
